// server/monitor/types.ts
export interface JaegerTrace {
    traceID: string;
    spans: JaegerSpan[];
    processes: Record<string, JaegerProcess>;
}

export interface JaegerSpan {
    traceID: string;
    spanID: string;
    operationName: string;
    references: Array<{
        refType: string;
        traceID: string;
        spanID: string;
    }>;
    startTime: number; // microseconds
    duration: number;  // microseconds
    tags: Array<{ key: string; type: string; value: any }>;
    processID: string;
}

export interface JaegerProcess {
    serviceName: string;
    tags: Array<{ key: string; type: string; value: any }>;
}